backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
